requirements.txt
pandas
numpy
openai
transformers
matplotlib
pyyaml
pytest
pymongo
langchain
langchain-community
chromadb
tiktoken
pyyaml
scikit-learn